monitoring/jaeger/python-app/requirements.txt
opentelemetry-api
opentelemetry-sdk
# opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp-proto-http
Flask